@@ -1,9 +1,9 @@
 {
   "name": "@seneca/allow",
-  "version": "0.2.1",
-  "main": "dist/allow.js",
+  "version": "1.0.0",
+  "main": "dist/Allow.js",
   "type": "commonjs",
-  "types": "dist/allow.d.ts",
+  "types": "dist/Allow.d.ts",
   "description": "Allow messages to proceed if they meet certain criteria",
   "homepage": "https://github.com/senecajs/seneca-allow",
   "keywords": [
@@ -16,16 +16,18 @@
     "url": "git://github.com/senecajs/seneca-allow.git"
   },
   "scripts": {
+    "prettier": "prettier --write .",
     "test": "jest --coverage",
     "test-some": "jest -t",
     "test-watch": "jest --coverage --watchAll",
     "watch": "tsc -w -d",
     "build": "tsc -d",
-    "clean": "rm -rf node_modules yarn.lock package-lock.json",
+    "doc": "seneca-doc",
+    "clean": "rm -rf node_modules yarn.lock package-lock.json dist",
     "reset": "npm run clean && npm i && npm test",
     "repo-tag": "REPO_VERSION=`node -e \"console.log(require('./package').version)\"` && echo TAG: v$REPO_VERSION && git commit -a -m v$REPO_VERSION && git push && git tag v$REPO_VERSION && git push --tags;",
     "repo-publish": "npm run clean && npm i && npm run repo-publish-quick",
-    "repo-publish-quick": "npm run build && npm run test && npm run repo-tag && npm publish --access public --registry https://registry.npmjs.org "
+    "repo-publish-quick": "npm run build && npm run test && npm run doc && npm run repo-tag && npm publish --access public --registry https://registry.npmjs.org "
   },
   "license": "MIT",
   "files": [
@@ -34,20 +36,21 @@
     "LICENSE"
   ],
   "devDependencies": {
-    "@types/jest": "^29.5.7",
-    "esbuild": "^0.19.5",
+    "@types/jest": "^29.5.11",
+    "esbuild": "^0.19.9",
     "esbuild-jest": "^0.5.0",
     "jest": "^29.7.0",
     "seneca": "^3.23.3",
     "seneca-msg-test": "^3.4.0",
     "seneca-promisify": "^2.0.0",
-    "typescript": "^5.2.2"
+    "seneca-doc": "^2.1.3",
+    "typescript": "^5.3.3"
   },
   "peerDependencies": {
     "seneca": ">=3",
     "seneca-promisify": ">=3"
   },
   "dependencies": {
-    "@hapi/hoek": "^11.0.2"
+    "@hapi/hoek": "^11.0.4"
   }
 }
